% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axisrange.R
\name{autorange_y}
\alias{autorange_y}
\title{autorange_y}
\usage{
autorange_y(fileName, Nrm = TRUE, nMajorTicks = 5, DfltZero = TRUE)
}
\arguments{
\item{fileName}{Input file name. Data should be arranged same as the input file for \code{\link{frogplots}}.Case sensitive and be sure to type with quotation marks. Currently only takes \code{.csv} files.}

\item{Nrm}{When \code{TRUE}, normalize data to control/first group (as 1). Default is \code{TRUE}.}

\item{nMajorTicks}{Number of major ticks intended to use for the plot.}

\item{DfltZero}{When \code{TRUE}, start y axis from \code{0}. Default is \code{TRUE}.}
}
\value{
A list object containing \code{lower_limit}, \code{upper_limit}, \code{major_tick_range} and \code{minor_tick_options}.
}
\description{
A function to get custom lower/upper limit, major tick range, as well as minor tick options for y axis, based on a user-defined major tick number.
}
\examples{
\dontrun{
autorange_y("data.csv",Nrm = T, nMajorTicks = 8, DfltZero=FALSE)
}
}

