% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frogplots.R
\name{frogplots}
\alias{frogplots}
\title{frogplots}
\usage{
frogplots(fileName, Tp = "Tukey", xAngle = 0, xAlign = 0.5,
  Title = NULL, xLabel = NULL, yLabel = NULL, legendTtl = FALSE,
  plotWidth = 170, plotHeight = 150)
}
\arguments{
\item{fileName}{Input file name. Case sensitive and be sure to type with quotation marks. Currently only takes \code{.csv} files.}

\item{Tp}{Type of the intended statistical test. Case sensitive and be sure to type with quotation marks. Options are: "t-test", "ANOVA", "Tukey" and "Dunnett". Default is "Tukey".}

\item{xAngle}{The rotation angle (degrees) of the x axis marks. Default is \code{0} - horizontal.}

\item{xAlign}{The alignment type of the x axis marks. Options are \code{0}, \code{0.5} and \code{1}. The default value at \code{0} is especially useful when \code{xAngle = 90}.}

\item{Title}{The displayed title on top of the plot. Be sure to type with quotation marks. Default is \code{NULL}.}

\item{xLabel}{x axis label. Type with quotation marks. Default is \code{NULL}}

\item{yLabel}{y axis label. Type with quotation marks. Default is \code{NULL}}

\item{legendTtl}{Hide/Display legend title. If \code{TRUE} or \code{T}, the name of the first column of the raw date file will display as the legend title. Default is \code{FALSE}.}

\item{plotWidth}{The width of the plot (unit: mm). Default is 170. Default will fit most of the cases.}

\item{plotHeight}{The height of the plot (unit: mm). Default is 150. Default will fit most of the cases.}
}
\value{
Outputs a \code{.csv} file with detailed metrics for the plot, including Mean, SEM and significance labels, as well as a plot image file (\code{.pdf}), with 600 dpi resolution.
}
\description{
A simple to use function for plotting basing on the statistical analysis of choice.
}
\examples{
\dontrun{
frogplots("data.csv",Tp="Tukey",yLabel="Relative fluorescence level")
frogplots("data2.csv",Tp="t-test",xAngle = -90, xAlign=0,yLabel="Relative fluorescence level")
frogplots("data3.csv",Tp="Tukey",yLabel="Relative fluorescence level")
frogplots("data4.csv",Tp="Dunnett", yLabel="Relative fluorescence level")
frogplots("data5.csv",Tp="Tukey",yLabel="Relative fluorescence level", plotWidth = 300)
}
}

