% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frogstats.R
\name{rbiostats}
\alias{rbiostats}
\title{rbiostats}
\usage{
rbiostats(fileName, Tp = "ANOVA")
}
\arguments{
\item{fileName}{Input file name. Case sensitive and be sure to type with quotation marks. Currently only takes \code{.csv} files.}

\item{Tp}{Type of the intended statistical test. Case sensitive and type with quotation marks. Options are: "t-test", "ANOVA", "Tukey" and "Dunnett". Default is "ANOVA".}
}
\value{
Outputs a \code{.txt} file with Shapiro-Wilk normality test results and the results of the statistical analysis of interest.
}
\description{
A simple to use function for comprehensive statistical analyses.
}
\examples{
\dontrun{
rbiostats("data.csv","Tukey")
rbiostats("data2.csv","t-test")
}
}

