% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axisrange.R
\name{all_dvsr}
\alias{all_dvsr}
\title{all_dvsr}
\usage{
all_dvsr(x, i = 0)
}
\arguments{
\item{x}{The number of interest}

\item{i}{Decimal scaling factor. If the input number is an integer, the starting number will be \code{x*10^i}; if a decimal, the starting number will be \code{x*10^n*10^i} (\code{x*10^(n+i)}), where the auto-determined \code{n} is the decimal amplification factor that transforms the decimal numbers to their corresponding minimum intergers by \code{10*n}.}
}
\value{
A vector containing all the divisors for the input number
}
\description{
A function to find all divisors for any types of number
}
\examples{
\dontrun{
all_dvsr(0.5)
all_dvsr(0.6,1) # with decimal scaling factor set at 1, meaning the divisor will be based on 60.
all_dvsr(2)
all_dvsr(2,1) # with decimal scaling factor set at 1, meaning the divisor will be based on 20.
}
}

