% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frogplots.R
\name{frogplots}
\alias{frogplots}
\title{frogplots}
\usage{
frogplots(fileName, Tp = "Tukey", xAngle = 0, xAlign = 0.5,
  Title = NULL, xLabel = NULL, yLabel = NULL, legendTtl = FALSE,
  plotWidth = 170, plotHeight = 150, y_custom_tick_range = FALSE,
  y_lower_limit = 0, y_upper_limit, y_major_tick_range, y_n_minor_ticks = 4)
}
\arguments{
\item{fileName}{Input file name. Case sensitive and be sure to type with quotation marks. Currently only takes \code{.csv} files.}

\item{Tp}{Type of the intended statistical test. Case sensitive and be sure to type with quotation marks. Options are: "t-test", "ANOVA", "Tukey" and "Dunnett". Default is "Tukey".}

\item{xAngle}{The rotation angle (degrees) of the x axis marks. Default is \code{0} - horizontal.}

\item{xAlign}{The alignment type of the x axis marks. Options are \code{0}, \code{0.5} and \code{1}. The default value at \code{0} is especially useful when \code{xAngle = 90}.}

\item{Title}{The displayed title on top of the plot. Be sure to type with quotation marks. Default is \code{NULL}.}

\item{xLabel}{x axis label. Type with quotation marks. Default is \code{NULL}.}

\item{yLabel}{y axis label. Type with quotation marks. Default is \code{NULL}.}

\item{legendTtl}{Hide/Display legend title. If \code{TRUE} or \code{T}, the name of the first column of the raw date file will display as the legend title. Default is \code{FALSE}.}

\item{plotWidth}{The width of the plot (unit: mm). Default is 170. Default will fit most of the cases.}

\item{plotHeight}{The height of the plot (unit: mm). Default is 150. Default will fit most of the cases.}

\item{y_custom_tick_range}{To initiate setting custom \code{y_upper_limit}, \code{y_lower_limit}, \code{y_major_tick_range}, \code{y_n_minor_ticks}. Default is \code{FALSE}.}

\item{y_lower_limit}{Can only be set when \code{y_custom_tick_range = TRUE}. Set custom lower limt for y axis. Default is \code{0}. Value can be obtained from \code{\link{frogrange}}.}

\item{y_upper_limit}{Can only be set when \code{y_custom_tick_range = TRUE}. Set custom upper limt for y axis. Value can be obtained from \code{\link{frogrange}}.}

\item{y_major_tick_range}{Can only be set when \code{y_custom_tick_range = TRUE}. Set custom major tick range for y axis.  Value can be obtained from \code{\link{frogrange}}.}

\item{y_n_minor_ticks}{Can only be set when \code{y_custom_tick_range = TRUE}. Set custom numbers of minor ticks. Default is \code{4}. Value can be obtained from \code{\link{frogrange}}.}
}
\value{
Outputs a \code{.csv} file with detailed metrics for the plot, including Mean, SEM and significance labels, as well as a plot image file (\code{.pdf}), with 600 dpi resolution.
}
\description{
A simple to use function for plotting basing on the statistical analysis of choice.
}
\examples{
\dontrun{
frogplots("data.csv", Tp = "Tukey",
yLabel = "Relative fluorescence level")

frogplots("data2.csv", Tp = "t-test", xAngle = -90,
xAlign=0,yLabel="Relative fluorescence level")

frogplots("data3.csv", Tp = "Tukey",
yLabel = "Relative fluorescence level")

frogplots("data4.csv", Tp = "Dunnett",
yLabel = "Relative fluorescence level")

frogplots("data5.csv", Tp = "Tukey",
yLabel = "Relative fluorescence level", plotWidth = 300)

frogplots("data8.csv", Tp = "Tukey",
yLabel = "Relative fluorescence level",
y_custom_tick_range = TRUE,y_upper_limit = 4,
y_lower_limit = 0, y_major_tick_range = 1,
y_n_minor_ticks = 4)
}
}

